site: Add extra CSS for links
Switching to purple!  This breaks with the generally accepted "blue is
unvisited and purple is visited" paradigm, but it fits much better with
the general theme of the site.

Other ideas would be to crib some blue colour from modus-themes, or turn
everything gey-ish (something like #333, but that would perhaps be hard
to read for people just scrolling through looking for links).