css: Add responsive table of contents
If the screen is big enough to support a left-side margin (and it is,
except on mobile), insert a table of contents that moves with the
viewport.  If not, fallback to the old-style table of contents that we
had before.  Also, make the links less obtrusive inside of the TOC by
colouring the black, yet mark them as links by using small caps.