scripts/maths.js: Output only HTML
KaTeX outputs invalid MathML sometimes, see e.g. [1] and [2], which
results in invalid HTML. Tools such as [temml] apparently have a much
better MathML output, but browser support for the format as a whole is
still… [wonky]. As such, it seems best to get rid of MathML for now, and
reevaluate its importance at a later date.

[1]: https://github.com/KaTeX/KaTeX/issues/4078
[2]: https://github.com/KaTeX/KaTeX/issues/3978
[temml]: https://temml.org/
[wonky]: https://temml.org/docs/en/administration#browser-issues